backend/pyproject.toml
[project]
name = "restaurant-booking-agent-backend"
version = "0.1.0"
description = "FastAPI + Strands Agents backend for the restaurant booking assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "uvicorn>=0.30.0",
    "strands-agents[otel]>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.34.0",
    "sst",
    "sse-starlette>=3.2.0",
    "aws-lambda-powertools[tracer]>=3.25.0",
    "braintrust[otel]>=0.0.176",
]

[dependency-groups]
dev = [
    "strands-agents-evals>=0.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "moto[dynamodb]>=5.0.0",
    "ty>=0.0.1",
    "ruff>=0.9.0",
    "pylint>=4.0.5",
    "autoevals>=0.0.99",
    "python-dotenv>=1.0.0",
]

# SST Python SDK is not published to PyPI — install from source
[tool.uv.sources]
sst = { git = "https://github.com/sst/sst.git", subdirectory = "sdk/python", branch = "dev" }

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app", "evals"]

[tool.pylint.messages_control]
disable = [
    "C0301",  # line-too-long — ruff enforces line length; avoid double-reporting
    "C0411",  # wrong-import-order — ruff/isort owns import ordering
    "R0801",  # duplicate-code — false positive on structurally similar Pydantic field lists
    "R0903",  # too-few-public-methods — Starlette middleware only requires dispatch()
]

[tool.pylint.format]
max-line-length = 88  # match ruff

[tool.pytest.ini_options]
testpaths = ["tests", "evals"]
asyncio_mode = "auto"
markers = [
    "integration: requires a deployed AWS test stage — set INTEGRATION_TABLE_NAME / INTEGRATION_API_URL / INTEGRATION_CHAT_URL",
    "agent: requires real Bedrock credentials — run manually or nightly, not on every push",
]
